refactored and added portal class
