+++ .ruby-version
@@ -1,1 +1,1 @@
-2.1.0
+2.2.3
