+++ lib/bullet.rb
@@ -10,19 +10,19 @@ attr_accessor :x, :y, :angulo
 	@angulo = angulo
 	@imagens = Gosu::Image.new(@window,"media/bulletz3.png", true)
 	end
-	
+
 	def draw
 		imagem= @imagens
 		imagem.draw_rot(@x, @y, 4, @angulo)
 	end
-	
+
 	def move
 		self.x += Gosu::offset_x(@angulo, 20)
 		self.y += Gosu::offset_y(@angulo, 20)
 	end
-	
-	end
-	
+
+end
+
 class Clip
 	attr_reader :x, :y
 	attr_accessor :x, :y
@@ -34,13 +34,13 @@ class Clip
 		@imagens = Gosu::Image.new(@window,"media/bullets.png", true)
 	end
 
-	
+
 	def draw
 		imagem= @imagens
 		imagem.draw_rot(@x, @y, 2, (@angulo += 1)%360)
 	end
-	
-end	
+
+end
 
 class Shotgun
 	attr_reader :x, :y
@@ -53,12 +53,12 @@ class Shotgun
 		@imagens = Gosu::Image.new(@window, "media/m3_d.bmp", true)
 	end
 
-	
+
 	def draw
 		imagem= @imagens
 		imagem.draw_rot(@x, @y, 2, @angulo)
 	end
-	
-end	
+
+end
